requirements.txt
keytotext==2.3.2
opencv-python==4.6.0.66
protobuf==3.20.*
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
ipapy==0.0.9.0
pandas==1.5.0
positional-encodings==6.0.1
pytorch-lightning==1.6.4
redis==4.3.4
dragonmapper==0.2.6
chinese-english-lookup==0.0.3
docarray==0.21.0
streamlit==1.17.0
altair==4.0